+++ R/helpers.R
@@ -1,3 +1,80 @@
+# Forces evaluation of all functions
+# In order to produce referential transparency and ensure the sub-arguments
+# don't get copied over
+# see similar function in the scales package
+force_all <- function(...) {
+  list(...)
+}
+
+#' Function-returning wrapper for format
+#'
+#' This function takes a set of options for the \code{format()} function and returns a function that itself calls \code{format()} with those settings.
+#'
+#' The only differences are:
+#'
+#' 1. \code{scientific} is set to \code{FALSE} by default, and \code{trim} is set to \code{TRUE}
+#' 2. Passing a \code{NA} value produces \code{''} instead of \code{'NA'}.
+#' 3. In addition to standard \code{format()} options, it also accepts a \code{percent} option to apply percentage formatting, and \code{prefix} and \code{suffix} options to apply prefixes or suffixes to formatted numbers.
+#' 4. Has an attribute \code{'big.mark'} storing the \code{'big.mark'} option chosen.
+#'
+#' This is in the spirit of the \code{label_} functions in the scales package, except that it uses \code{format()}'s focus on significant digits instead of fixed decimal places, which is good for numbers that range across multiple orders of magnitude, common in \code{sumtable()} and \code{vtable()}.
+#'
+#' @param percent Whether to apply percentage formatting. Set to \code{TRUE} if 1 = 100\%. Or, optionally, set to any other number that represents 100\%. So \code{percent = TRUE} or \code{percent = 1} will interpret \code{.9} as \code{90\%}, or \code{percent = 100} will format \code{90} as \code{90\%}.
+#' @param prefix A prefix to apply to the formatted number. For example, \code{prefix = '$'} would format \code{4} as \code{$4}.
+#' @param suffix A suffix to apply to the formatted number. If specified alongside \code{percent}, the suffix comes after the \%.
+#' @param digits Number of significant digits.
+#' @param nsmall The minimum number of digits to the right of the decimal point.
+#' @param big.mark A character to mark thousands places, for example producing "1,000" instead of "1000".
+#' @param trim Whether numbers should be trimmed to their own size, rather than being right-justified to a common width. Unlike the actual \code{format()}, this defaults to \code{TRUE}. Note that in most vtable applications, the formatting function is applied one value at a time, rather than to a vector, so \code{trim = FALSE} may not work as intended.
+#' @param scientific Whether numbers should be encoded in scientific format. Unlike the actual \code{format()}, this defaults to \code{FALSE}.
+#' @param ... Arguments to be passed to \code{format()}. See \code{help(format)}. All other parameters listed above except for \code{percent}, \code{prefix}, or \code{suffix} are also just part of  \code{format}, but may be of particular interest, or have been included to show how defaults have changed.
+#' @examples
+#' x <- c(1, 1000, .000235, 1298.255, NA)
+#' my.formatting.func = format.func(digits = 3, prefix = '$')
+#' my.formatting.func(x)
+#'
+#' @export format.func
+format.func <- function(percent = FALSE, prefix = '', suffix = '',
+                        digits = NULL, nsmall = 0L, big.mark = '',
+                        trim = TRUE, scientific = FALSE, ...) {
+  if (!is.null(digits)) {
+    if (is.na(digits)) {
+      digits <- NULL
+    }
+  }
+  scalefactor <- 1
+  if (is.numeric(percent)) {
+    scalefactor <- 1/percent
+    percent <- TRUE
+  }
+  if (percent) {
+    scalefactor <- scalefactor*100
+    suffix <- paste0('%',suffix)
+  }
+  force_all(scalefactor, digits, nsmall, big.mark, trim, scientific, prefix, suffix, ...)
+  the_function <- function(x) {
+    x_fmt <- format(x*scalefactor, digits = digits, nsmall = nsmall, big.mark = big.mark,
+                    trim = trim, scientific = scientific, ...)
+    if (prefix != '' & trim == FALSE) {
+      # If we need a prefix but we have blank-space padding at the start, add the prefix after
+      first_nonspace <- regexpr('\\S',x_fmt)
+      x_fmt <- sapply(1:length(x_fmt),
+                      function(i) paste0(
+                        c(rep(' ', first_nonspace[i]-1),
+                          prefix,
+                          substr(x_fmt[i],first_nonspace[i],nchar(x_fmt[i]))),
+                        suffix, collapse = ''))
+    } else {
+      x_fmt <- paste0(prefix, x_fmt, suffix)
+    }
+    x_fmt[is.na(x)] <- ''
+    return(x_fmt)
+  }
+  attr(the_function, 'big.mark') <- big.mark
+  return(the_function)
+}
+
+
 #' Number of unique values in a vector
 #'
 #' This function takes a vector and returns the number of unique values in that vector.
@@ -26,7 +103,7 @@ nuniq <- function(x) {
 #' w <- c(4, 1, 3, 7, 0, 2, 5)
 #' weighted.sd(x, w)
 #'
-#' @export
+#' @export weighted.sd
 weighted.sd <- function(x, w, na.rm = TRUE) {
   if (length(x) != length(w)) {
     stop('Weights and data must be the same length.')
@@ -88,13 +165,23 @@ countNA <- function(x) {
 #'
 #' This function just shorthand for \code{sum(!is.na(x))}, with a shorter name for reference in the \code{vtable} or \code{sumtable} \code{summ} option.
 #'
+#' If \code{big.mark} is specified, will return a formatted string instead of a number, where the formatting is based on \code{format(x, big.mark = big.mark, scientific = FALSE, ...)}.
+#'
 #' @param x A vector.
+#' @param big.mark Argument to pass to \code{format()}, if a formatted string is desired.
+#' @param scientific Argument to pass to \code{format()} if \code{big.mark} is specified. Defaults to \code{FALSE}, unlike in \code{format()}.
+#' @param ... Other arguments to pass to \code{format()}. Ignored if \code{big.mark} is not specified.
 #' @examples
 #' x <- c(1, 1, NA, 2, 3, NA)
 #' notNA(x)
+#' notNA(1:10000, big.mark = ',')
 #' @export
-notNA <- function(x) {
-  sum(!is.na(x))
+notNA <- function(x, big.mark = NULL, scientific = FALSE, ...) {
+  if (is.null(big.mark)) {
+    return(sum(!is.na(x)))
+  } else {
+    return(format(sum(!is.na(x)), big.mark = big.mark, scientific = scientific, ...))
+  }
 }
 
 #' Returns a vector of 100 percentiles
@@ -190,7 +277,7 @@ parsefcn_summ <- function(x,y, ...) {
 summary.row <- function(data,var,st,
                         title,summ,cla,factor.percent,
                         factor.count,factor.numeric,digits,fixed.digits, wts = NULL,
-                        fmt = NULL, skip.format = NULL) {
+                        fmt = NULL, skip.format = NULL, factor.not.numeric.count = notNA) {
 
   numcols <- length(summ)
   if (cla == 'header') {
@@ -201,7 +288,8 @@ summary.row <- function(data,var,st,
     #Get data
     va <- data[[var]]
     #Total number of obs
-    nonmiss <- notNA(va)
+    nonmiss <- factor.not.numeric.count(va)
+    nonmissnum <- notNA(va)
     #Header row
     st[1,] <- c(title,nonmiss,
                 #Take this out for now.
@@ -210,22 +298,20 @@ summary.row <- function(data,var,st,
                 rep('',numcols-2))
     #And now the per-factor stuff
     mat <- as.data.frame(table(va))
-    propcalc <- mat$Freq/nonmiss
+    matlabel <- aggregate(y~x, data.frame(y = 1, x = va), FUN = factor.not.numeric.count, drop = FALSE)
+    matlabel$y[is.na(matlabel$y)] <- 0
+    propcalc <- mat$Freq/nonmissnum
     if (!is.null(wts) & grepl('wts',summ[2])) {
       propcalc <- sapply(mat$va, function(x) stats::weighted.mean(va == x, w = wts, na.rm = TRUE))
     }
     propcalc <- propcalc*(100^factor.percent)
     mat$va <- paste('...',mat$va)
+    mat$Freq <- matlabel$y
     if (fixed.digits) {
       mat$Prop <- sapply(1:length(propcalc), function(x)
         format(propcalc[x],
                digits=max(digits[2]-2*factor.percent,1),
                nsmall=max(digits[2]-2*factor.percent,0),
-               scientific = FALSE))
-      mat$Freq <- sapply(1:length(mat$Freq), function(x)
-        format(as.numeric(mat$Freq[x]),
-               digits=max(digits[1],1),
-               nsmall=digits[1],
                scientific = FALSE))
       st[1,2] <- format(as.numeric(st[1,2]),
                                    max(digits = digits[1],1),
@@ -257,18 +343,18 @@ summary.row <- function(data,var,st,
     facnames <- paste('...',levels(va))
     #Run each of the functions on the variable and get results
     results <- lapply(1:ncol(mat),
-                      function(x) sapply(summ, function(y) parsefcn_summ(mat[,x],y, wts = wts[!is.na(va)])))
+                      function(x) lapply(summ, function(y) parsefcn_summ(mat[,x],y, wts = wts[!is.na(va)])))
     #Format
     results <- lapply(1:length(results),
-                      function(x) sapply(1:length(results[[x]]), function(y) ifelse(summ[y] %in% skip.format, results[[x]][y], fmt(results[[x]][y]))))
+                      function(x) sapply(1:length(results[[x]]), function(y) ifelse(summ[y] %in% skip.format, results[[x]][[y]], fmt(results[[x]][[y]]))))
     #Round
-    if (is.numeric(results[[1]])) {
+    if (paste0(deparse(fmt), collapse = '') == 'function (x) x') {
       if (fixed.digits) {
         results <- lapply(results, function(x)
-          sapply(1:length(x), function(y) format(x[y],digits=digits[y],nsmall = max(digits[y],1),scientific=FALSE)))
+          sapply(1:length(x), function(y) ifelse(is.character(x[y]), x[y], format(x[y],digits=digits[y],nsmall = max(digits[y],1),scientific=FALSE))))
       } else {
         results <- lapply(results, function(x)
-          sapply(1:length(x), function(y) as.character(round(x[y],digits=digits[y]))))
+          sapply(1:length(x), function(y) ifelse(is.character(x[y]), x[y], as.character(round(x[y],digits=digits[y])))))
       }
     }
     #Add factor names
@@ -282,17 +368,17 @@ summary.row <- function(data,var,st,
     va <- data[[var]]
 
     #Run each of the functions on the variable and get results
-    results <- sapply(summ, function(y) parsefcn_summ(va,y, wts = wts[!is.na(va)]))
+    results <- lapply(summ, function(y) parsefcn_summ(va,y, wts = wts[!is.na(va)]))
 
     #Format
-    results <- sapply(1:length(results), function(y) ifelse(summ[y] %in% skip.format, results[y], fmt(results[y])))
+    results <- lapply(1:length(results), function(y) ifelse(summ[y] %in% skip.format, results[[y]], fmt(results[[y]])))
 
-    # If it's still a number, our formatting function was function(x). If not, we did formatting
-    if (is.numeric(results)) {
+    # If our formatting function was function(x) x, apply digits options
+    if (paste0(deparse(fmt), collapse = '') == 'function (x) x') {
       if (fixed.digits) {
-        results <- sapply(1:length(results), function(y) format(results[y],digits=max(digits[y],1),nsmall = digits[y],scientific = FALSE))
+        results <- sapply(1:length(results), function(y) ifelse(is.character(results[[y]]), results[[y]], format(results[[y]],digits=max(digits[y],1),nsmall = digits[y],scientific = FALSE)))
       } else {
-        results <- sapply(1:length(results), function(y) round(results[y],digits=max(digits[y],1)))
+        results <- sapply(1:length(results), function(y) ifelse(is.character(results[[y]]), results[[y]], round(results[[y]],digits=max(digits[y],1))))
       }
     }
     #And construct
@@ -389,14 +475,14 @@ clean_multicol_kable <- function(df,title,note=NA) {
   }
 
   if (is.null(fmt)) {
-    if (!is.null(hasheader)) {
+    if (hasheader) {
       headerrow <- gsub('_MULTICOL.*$','',headerrow)
       names(headerrow) = names(df)
       df = rbind(headerrow, df)
     }
     kb <- knitr::kable(df, caption = title, row.names = FALSE)
   } else if (fmt == 'html') {
-    kb <- knitr::kable(df, caption = title, row.names = FALSE, format = fmt, escape = FALSE)
+    kb <- knitr::kable(df, caption = title, row.names = FALSE, format = fmt, escape = TRUE)
   } else if (fmt == 'latex') {
     kb <- knitr::kable(df, caption = title, row.names = FALSE, format = fmt, booktabs = TRUE, escape = FALSE)
   }
@@ -421,3 +507,5 @@ clean_multicol_kable <- function(df,title,note=NA) {
 
   return(kb)
 }
+
+